+++ run_tox.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 cd utils-test-app
-tox -e py36-django31
+tox -e py37-django32
 cd ..